requirements.txt
discord.py==2.3.2
Flask==2.0.1
python-dotenv==1.0.0
requests==2.28.1
